pyproject.toml
[project]
name = "human3090"
version = "0.0.1"
requires-python = ">=3.12"
authors = [{ name = "Mihai Cosma", email = "mcosma@gmail.com" }]
dependencies = [
    "dotenv>=0.9.9",
    "huggingface-hub>=0.30.2",
    "numpy>=1.26",
    "tqdm>=4.66",
    "matplotlib>=3.10.1",
    "openai>=1.72.0",
    "pyyaml>=6.0",
    "together>=1.5.5",
]

[project.scripts]
bench_runner = "human3090.bench_runner:main"
start_human_daemon = "human3090.daemon:start"
stop_human_daemon = "human3090.daemon:stop"
inspect_result = "human3090.inspect_result:main"
show_human = "human3090.show_human:main"
test_single_problem = "human3090.test_single_problem:main"

[tool.ruff]
# Default is: pycodestyle (E) and Pyflakes (F)
# We add flake8-builtins (A), pydocstyle (D), isort (I), pep8-naming (N), and pylint (PL).
# We remove pycodestyle (E) since it throws erroneous line too long errors.
# We remove Pyflakes (F) since it complains about `import *` which we need.
select = ["A", "D", "I", "N", "PL"]

# We ignore the following rules:
# A002: Argument `X` is shadowing a Python builtin
# D100: missing docstring in public module
# D101: missing docstring in public class
# D102: missing docstring in public method
# D103: missing docstring in public function
# D105: missing docstring in magic method
# D107: Missing docstring in `__init__`
# D203: 1 blank line required before class docstring (incompatible with D211: no blank lines before class docstring)
# D213: multi-line-summary-second-line (incompatible with D212: multi-line summary should start at the first line)
# D413: Missing blank line after last section
# D415: First line should end with a period, question mark, or exclamation point (in period-only D400)
# D416: section-name-ends-in-colon (numpy style guide doesn't use colons after sections, i.e. Parameters)
# PLR0911: Too many return statements (X > 6)
# PLR0912: Too many branches (X > 12)
# PLR0913: Too many arguments to function call (X > 5)
# PLR0915: Too many statements (X > 50)
# PLR2004: Magic value used in comparison
ignore = ["A002", "D100", "D101", "D102", "D103", "D105", "D107", "D203", "D213", "D413", "D415", "D416", "PLR0911", "PLR0912", "PLR0913", "PLR0915", "PLR2004"]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "D", "I", "N", "PL"]
unfixable = []

line-length = 200

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.12
target-version = "py312"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
"human3090.human_eval" = ["data/*.gz"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
